pyproject.toml
[tool.poetry]
name = "deepscapy"
version = "0.1.0"
description = ""
authors = ["Varun Nandkumar Golani <vngolani@mail.uni-paderborn.de>"]

[tool.poetry.dependencies]
python = "3.9.2"
numpy = "1.22.4"
scikit-learn = "1.1.1"
h5py = ">=3.1.0"
pandas = ">=1.1.5"
matplotlib = ">=3.3.3"
future = "0.18.2"
openpyxl = ">=3.0.0"
tensorflow = "2.9.0"
tensorflow_addons = "0.17.0"
autokeras = "1.0.19"
keras_tuner = "1.1.2"
protobuf = "==3.20.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
